LayeredArrays

Introduction

LayeredArrays provides layer-wise array computation written in the Julia programming language. The layers have hierarchical structure, and lower layers can be accessed by using getindex in AbstractLayeredArray. All types except subtypes of AbstractLayeredArray are on bottom layer 0. The layer-wise operations are simply available by using broadcast operations, where operations are performed in order from the highest to the lowest layer. For example, we have three vectors on layer 1 and 2:

julia> x = @layered 1 ["a", "b", "c"]
3-element LayeredVector{1, String, Vector{String}}:
 "a"
 "b"
 "c"

julia> y = @layered 1 ["d", "e", "f"]
3-element LayeredVector{1, String, Vector{String}}:
 "d"
 "e"
 "f"

julia> z = @layered 2 ["g", "h", "i"]
3-element LayeredVector{2, String, Vector{String}}:
 "g"
 "h"
 "i"

The broadcasting vector multiplication for those vectors are then computed as

julia> @. x * y # equal to built-in Array for operations on the same layers
3-element LazyLayeredVector{1, String, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{LayeredVector{1, String, Vector{String}}, LayeredVector{1, String, Vector{String}}}}}:
 "ad"
 "be"
 "cf"

julia> @. x * z # broadcasting operations on each layer
3-element LazyLayeredVector{2, LazyLayeredVector{1, String, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{LayeredVector{1, String, Vector{String}}, Base.RefValue{String}}}}, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{Base.RefValue{LayeredVector{1, String, Vector{String}}}, LayeredVector{2, String, Vector{String}}}}}:
 ["ag", "bg", "cg"]
 ["ah", "bh", "ch"]
 ["ai", "bi", "ci"]

julia> @. x * y * z
3-element LazyLayeredVector{2, LazyLayeredVector{1, String, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{LayeredVector{1, String, Vector{String}}, LayeredVector{1, String, Vector{String}}, Base.RefValue{String}}}}, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{Base.RefValue{LayeredVector{1, String, Vector{String}}}, Base.RefValue{LayeredVector{1, String, Vector{String}}}, LayeredVector{2, String, Vector{String}}}}}:
 ["adg", "beg", "cfg"]
 ["adh", "beh", "cfh"]
 ["adi", "bei", "cfi"]

Note that the layer-wise broadcasting operations are always lazily evaluated.

Installation

pkg> add https://github.com/KeitaNakamura/LayeredArrays.jl.git

Types and Functions

LayeredArrays.AbstractLayeredArray — Type

AbstractLayeredArray{layer, T, N}

Supertype for N-dimensional layered arrays with elements of type T.

LayeredArrays.LayeredArray — Type

LayeredArray(x)
LayeredArray{layer}(x)

Construct LayeredArray on layer (layer = 1 by default).

See also @layered.

Examples

julia> x = LayeredArray([1,2,3])
3-element LayeredVector{1, Int64, Vector{Int64}}:
 1
 2
 3

julia> y = LayeredArray{2}([4,5,6])
3-element LayeredVector{2, Int64, Vector{Int64}}:
 4
 5
 6

julia> x .* y
3-element LazyLayeredVector{2, LazyLayeredVector{1, Int64, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{LayeredVector{1, Int64, Vector{Int64}}, Base.RefValue{Int64}}}}, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{Base.RefValue{LayeredVector{1, Int64, Vector{Int64}}}, LayeredVector{2, Int64, Vector{Int64}}}}}:
 [4, 8, 12]
 [5, 10, 15]
 [6, 12, 18]

LayeredArrays.@layered — Macro

@layered expr
@layered layer expr

Construct LayeredArray on layer (layere = 1 by default). This is equivalent to LayeredArray{layer}(expr).

See also LayeredArray.

Examples

julia> x = @layered [1,2,3]
3-element LayeredVector{1, Int64, Vector{Int64}}:
 1
 2
 3

julia> y = @layered 2 [4,5,6]
3-element LayeredVector{2, Int64, Vector{Int64}}:
 4
 5
 6

julia> x .* y
3-element LazyLayeredVector{2, LazyLayeredVector{1, Int64, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{LayeredVector{1, Int64, Vector{Int64}}, Base.RefValue{Int64}}}}, Base.Broadcast.Broadcasted{LayeredArrays.LayeredArrayStyle{1}, Nothing, typeof(*), Tuple{Base.RefValue{LayeredVector{1, Int64, Vector{Int64}}}, LayeredVector{2, Int64, Vector{Int64}}}}}:
 [4, 8, 12]
 [5, 10, 15]
 [6, 12, 18]